pass1: Online Dictionary Learning core — seed/accumulate/refresh (Step 3b)
New core::pass1 module. Turns a signature stream into K binary codewords
(spec §4) via seed → stream → batch-refresh, no incoherence yet.

- seed(): ridge-leverage sampling of warmup signatures off the FD factor
  (Efraimidis–Spirakis weighted reservoir); each seed atom = top-D
  surprisal-weighted bits of the sampled signature.
- observe(): α-threshold pursuit → accumulate the signature's surprisal-
  weighted bits into each fired atom (b_k), bump n_k; auto-refresh.
- refresh(): φ_k ← top-D of b_k (D cap by construction) + decayed forgetting.
- utilization() + mean_recall() diagnostics.

SPEC AUGMENTATION (approved): the spec's "shadow-float weights, hard binary
forward, freeze" is realized as the decayed surprisal-weighted accumulator
b_k (soft) → top-D (hard), avoiding an infeasible dense K×3F float matrix.

Also widen pursuit weights uint16→uint32: atom info sums fit uint16, but
full-signature reconstruction sums (thousands of bits) would overflow.

Tests: D-cap invariant, utilization, surprisal-weighted recall on 40
disjoint patterns, determinism under fixed seed.

+++ src/core/codebook/include/pursuit.hpp
@@ -34,12 +34,13 @@ struct PursuitConfig {
   std::size_t max_fired = 32;    // budget on |fired|
 };
 
-// weights: per-3F-bit surprisal multiplicities (WGT1), size == codebook.dim().
+// weights: per-3F-bit surprisal multiplicities (WGT1), widened to uint32 so
+// full-signature info sums can't overflow; size == codebook.dim().
 // signature: sorted-unique 3F set positions. Returns fired atom ids, sorted.
 std::vector<std::uint32_t> pursuit_encode(
     const Codebook& cb,
     const std::vector<std::uint32_t>& signature,
-    const std::vector<std::uint16_t>& weights,
+    const std::vector<std::uint32_t>& weights,
     const PursuitConfig& cfg = {});
 
 }  // namespace core::codebook
